Create infInt class v1.0
This is a c++ program of a class to help with the manipulation of big integers (larger than the 64-bit integer range). Currently, it only supports basic operators (not including bitwise operators). Also, the operators are not optimized, so do not expect fast runtime when using them.
Warning: Division currently uses binary search and is extremely slow. I will be working on faster operators in future versions.